@@ -1,8 +1,23 @@
 {
     "name": "docker-php/docker-php",
+    "description": "A Docker PHP client",
     "license": "MIT",
     "type": "library",
-    "description": "A Docker PHP client",
+    "require": {
+        "php": "^8.3",
+        "jane-php/open-api-runtime": "^7.10",
+        "php-http/discovery": "^1.20",
+        "symfony/filesystem": "^7.4 || ^8.0",
+        "symfony/http-client": "^8.0",
+        "symfony/process": "^7.4 || ^8.0"
+    },
+    "require-dev": {
+        "jane-php/open-api-2": "^7.10",
+        "phpunit/phpunit": "^12.0",
+        "sigwin/infra": "~1.24.0"
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "autoload": {
         "psr-4": {
             "Docker\\": "src/"
@@ -13,41 +28,14 @@
             "Docker\\Tests\\": "tests/"
         }
     },
-    "require": {
-        "php": ">=7.1",
-        "docker-php/docker-php-api": "4.1.*",
-        "guzzlehttp/psr7": "^1.2",
-        "php-http/client-common": "^1.6",
-        "php-http/socket-client": "^1.3",
-        "php-http/message": "^1.0",
-        "symfony/filesystem": "^2.3 || ^3.0 || ^4.0",
-        "symfony/process": "^2.3 || ^3.0 || ^4.0"
-    },
-    "suggest": {
-        "php-http/httplug-bundle": "For integration with Symfony",
-        "amphp/artax": "To use the async api"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^6.0",
-        "friendsofphp/php-cs-fixer": "2.8.1",
-        "amphp/artax": "^3.0",
-        "amphp/socket": "^0.10.5"
-    },
-    "conflict": {
-        "amphp/socket": "<0.10.5",
-        "amphp/artax": "<3.0"
-    },
-    "scripts": {
-        "test": "vendor/bin/phpunit",
-        "test-ci": "vendor/bin/phpunit --coverage-clover build/coverage.xml",
-        "lint": "vendor/bin/php-cs-fixer fix --dry-run --verbose --diff",
-        "lint-fix": "vendor/bin/php-cs-fixer fix --verbose"
+    "config": {
+        "allow-plugins": {
+            "php-http/discovery": false
+        }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.x-dev"
+            "dev-master": "3.x-dev"
         }
-    },
-    "prefer-stable": true,
-    "minimum-stability": "dev"
+    }
 }
