@@ -2,6 +2,17 @@
 
 declare(strict_types=1);
 
+/*
+ * This file is part of the docker-php project.
+ *
+ * (c) 2013 Geoffrey Bachelet <geoffrey.bachelet@gmail.com> and contributors
+ * (c) 2019 Joël Wurtz
+ * (c) 2026 sigwin.hr
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Docker\Context;
 
 /**
@@ -11,10 +22,8 @@ interface ContextInterface
 {
     /**
      * Whether the Context should be streamed or not.
-     *
-     * @return bool
      */
-    public function isStreamed();
+    public function isStreamed(): bool;
 
     /**
      * If `isStreamed()` is `true`, then `read()` should return a resource.
