@@ -2,6 +2,17 @@
 
 declare(strict_types=1);
 
+/*
+ * This file is part of the docker-php project.
+ *
+ * (c) 2013 Geoffrey Bachelet <geoffrey.bachelet@gmail.com> and contributors
+ * (c) 2019 Joël Wurtz
+ * (c) 2026 sigwin.hr
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
 namespace Docker\Stream;
 
 use Psr\Http\Message\StreamInterface;
@@ -19,8 +30,6 @@ abstract class CallbackStream
 
     /**
      * Called when there is a new frame from the stream.
-     *
-     * @param callable $onNewFrame
      */
     public function onFrame(callable $onNewFrame): void
     {
@@ -29,21 +38,19 @@ abstract class CallbackStream
 
     /**
      * Read a frame in the stream.
-     *
-     * @return mixed
      */
-    abstract protected function readFrame();
+    abstract protected function readFrame(): mixed;
 
     /**
      * Wait for stream to finish and call callables if defined.
      */
     public function wait(): void
     {
-        while (!$this->stream->eof()) {
+        while (! $this->stream->eof()) {
             $frame = $this->readFrame();
 
-            if (null !== $frame) {
-                if (!\is_array($frame)) {
+            if ($frame !== null) {
+                if (! \is_array($frame)) {
                     $frame = [$frame];
                 }
 
